Make inline svg-references unique
mathematical.parse always creates the same id, "glyph-0-0".  When
inlining in html this results in clashing svgs which are not
rendered correctly.

Fix by inserting a counter in the id, resulting in unique ids.

Changes ids from "glyph-0-0" to eg. "glyph-19-0-0".

Test-document test.adoc:

````
:stem:
latexmath:[A] latexmath:[B] latexmath:[C]
````

Run:

In a browser (chrome and firefox tested) this looks like "A A A" before
the commit, and "A B C" after.